pyproject.toml
[project]
name = "personal-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "markdown>=3.7",
    "jinja2>=3.1.4",
    "python-frontmatter>=1.1.0",
]
